Allow for moving and syncing account info
Relevant account information can now be synced from inter to the
account's graph.  This will help when we start the login process and it
allows us to add the relevant information to the session graph.  That
should in turn make it easier to render information on the current user
in the interface.