Add login flow
The new system has support for a full login flow.  This allows us to log
a user in based on their email address and key.  A bunch of this flow is
still hardcoded based on our current dev setup as we await external
credentials to use.  Next steps are making those configurable so we can
already start testing the service in a development setup.